Fix CI round 3: Linux socket-constant types + two review findings
Linux build (Glibc types differ from Darwin):
- SOCK_STREAM is a `__socket_type` enum on Glibc (Int32 on Darwin): add a
  portable `makeUnixStreamSocket()` in PlatformSys and use it in
  EndpointConnector + DaemonServer.
- SHUT_RDWR / SOMAXCONN are `Int` on Glibc: cast to Int32 at the shutdown()
  and listen() call sites.

Review findings (Cursor Bugbot):
- RealPty Linux PTY child: if opening the slave fails, `_exit(127)` instead of
  exec'ing the shell with no controlling terminal and the inherited master fd
  as stdio.
- HarnessPaths.tunnelSocketURL: distinct host names that sanitize to the same
  text (e.g. "dev.box" vs "dev-box") produced the same socket path, so spawning
  the second tunnel clobbered the first. Disambiguate with a deterministic
  FNV-1a hash of the full name (Hasher is per-process seeded, so unusable for a
  stable on-disk name).

+++ Packages/HarnessCore/Sources/HarnessCore/IPC/DaemonClient.swift
@@ -219,7 +219,7 @@ public final class DaemonSubscription: @unchecked Sendable {
         // shutdown — paired with the read loop setting `finished` under the same lock
         // before it closes — guarantees `fd` is still open here, so we never shutdown a
         // descriptor the loop already closed and the OS may have recycled.
-        shutdown(fd, SHUT_RDWR)
+        shutdown(fd, Int32(SHUT_RDWR)) // SHUT_RDWR is `Int` on Glibc, `Int32` on Darwin
     }
 
     /// Output-stream convenience: forwards `.data` frames to `onData`.
